+++ src/kubernetes/config_yaml.rs
@@ -1,6 +1,4 @@
-use sha2::{Digest, Sha256};
-
-use super::constants::TOR_AGABANI_CO_UK_CONFIG_HASH_KEY;
+use super::Annotation;
 
 pub struct ConfigYaml(String);
 
@@ -10,23 +8,9 @@ impl ConfigYaml {
     }
 }
 
-impl ConfigYaml {
-    #[must_use]
-    pub fn to_annotation_tuple(&self) -> (String, String) {
-        (TOR_AGABANI_CO_UK_CONFIG_HASH_KEY.into(), self.sha_256())
-    }
-
-    #[must_use]
-    pub fn sha_256(&self) -> String {
-        let mut sha = Sha256::new();
-        sha.update(&self.0);
-        format!("sha256:{:x}", sha.finalize())
-    }
-}
-
-impl From<ConfigYaml> for String {
-    fn from(value: ConfigYaml) -> Self {
-        value.0
+impl AsRef<str> for ConfigYaml {
+    fn as_ref(&self) -> &str {
+        self.0.as_str()
     }
 }
 
@@ -34,4 +18,8 @@ impl From<&ConfigYaml> for String {
     fn from(value: &ConfigYaml) -> Self {
         value.0.clone()
     }
+}
+
+impl Annotation for ConfigYaml {
+    const NAME: &'static str = "config-yaml";
 }