+++ src/kubernetes/hostname.rs
@@ -1,0 +1,7 @@
+use crate::crypto::Hostname;
+
+use super::annotation::Annotation;
+
+impl Annotation for Hostname {
+    const NAME: &'static str = "hostname";
+}