+++ src/onion_balance.rs
@@ -393,10 +393,9 @@ async fn reconciler(object: Arc<OnionBalance>, ctx: Arc<Context>) -> Result<Acti
 
     let namespace = object.try_namespace()?;
 
-    let torrc = generate_torrc(&object);
     let config_yaml = generate_config_yaml(&object);
+    let torrc = generate_torrc(&object);
 
-    let annotations = generate_annotations(&config_yaml, &torrc);
     let labels = object.try_labels()?;
     let selector_labels = object.try_selector_labels()?;
 
@@ -408,6 +407,11 @@ async fn reconciler(object: Arc<OnionBalance>, ctx: Arc<Context>) -> Result<Acti
     .await?;
 
     if let State::Initialized(onion_key) = &state {
+        let annotations = Annotations::new()
+            .add(&config_yaml)
+            .add_opt(&onion_key.hostname())
+            .add(&torrc);
+
         // ConfigMap
         reconcile_config_map(
             &Api::new(kube::Api::namespaced(ctx.client.clone(), &namespace)),
@@ -523,7 +527,7 @@ async fn reconcile_onion_balance(
                 .unwrap_or(&Vec::new())
                 .merge_from(&state.into()),
             hostname: if let State::Initialized(onion_key) = state {
-                onion_key.hostname().map(Into::into)
+                onion_key.hostname().as_ref().map(ToString::to_string)
             } else {
                 None
             },
@@ -531,13 +535,6 @@ async fn reconcile_onion_balance(
         },
     )
     .await
-}
-
-fn generate_annotations(config_yaml: &ConfigYaml, torrc: &Torrc) -> Annotations {
-    Annotations::new(BTreeMap::from([
-        config_yaml.to_annotation_tuple(),
-        torrc.to_annotation_tuple(),
-    ]))
 }
 
 #[allow(unused_variables)]